requirements.txt
pdfplumber
groq
pandas
tqdm
sentence-transformers
faiss-cpu
numpy
regex
python-dotenv
